requirements.txt
# requirements.txt - Complete with advanced features
PyMuPDF>=1.23.0
pdfplumber>=0.9.0
pandas>=2.0.0
numpy>=1.21.0
scikit-learn>=1.2.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
textblob>=0.17.0
spacy>=3.4.0
wordcloud>=1.8.0
nltk>=3.7.0
vaderSentiment>=3.3.0
streamlit>=1.20.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
watchdog>=3.0.0
pytesseract>=0.3.10
Pillow>=9.0.0
transformers>=4.20.0
torch>=1.12.0
sentence-transformers>=2.2.0
umap-learn>=0.5.0
hdbscan>=0.8.0
networkx>=2.8.0
scipy>=1.9.0
kaleido>=0.2.1
